% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chi2_for_crosstabs.R
\name{get_chi2_for_crosstabs}
\alias{get_chi2_for_crosstabs}
\title{CHi-squared statis for crosstabulations}
\usage{
get_chi2_for_crosstabs(tables, strict = F, prefix = "n_")
}
\arguments{
\item{tables}{dataframe/tibble; a set of crosstabulations created with \code{get_crosstabs()} or \code{get_weighted_crosstabs()}}

\item{strict}{logical; If true, tables with <5 observations in a cell, get marked as "not enough observations" if False, number of observations <30 per table will produce this mark}

\item{prefix}{character; prefix of observation count columns in a set of tables; "n_"}
}
\value{
dataframe; set of tables with columns (see \code{chisq.test()} for statistical test details)
\itemize{
\item chisq_stat - chi-squared statistics;
\item chisq_pval - pvalue for chi-squared;
\item hyp - variables' independence;
\item t_assumption_n - sufficiency of observations for test validity;
}
}
\description{
CHi-squared statis for crosstabulations
}
