% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cleaning_log.R
\name{get_cleaning_log}
\alias{get_cleaning_log}
\title{Cleaning log (value and variable change)}
\usage{
get_cleaning_log(raw, clean, idvar = "_uuid", moda = F)
}
\arguments{
\item{raw}{dataframe; raw/original dataset}

\item{clean}{dataframe; clean/processed dataset}

\item{idvar}{character; variable name for unique observation IDs (e.g. "uuid")}

\item{moda}{logical; in "select multiple" substitutions of TRUE/FALSE to 0/1 is not tracked}
}
\value{
list containing 3 dataframes;
\itemize{
\item deletion - observations deleted from a clean dataset
\item variable change - indication of variables, present in one of datasets only
\item value cleaning - changes in variable's values between raw and clean dataset
}
}
\description{
Cleaning log (value and variable change)
}
