% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codebook_from_name_label.R
\name{get_codebook_from_name_label}
\alias{get_codebook_from_name_label}
\title{Get a codebook from a dataset with a double header}
\usage{
get_codebook_from_name_label(ds)
}
\arguments{
\item{ds}{dataframe; a dataset, where the header is xls name, while the first row is the xls label}
}
\value{
dataframe; a codebook with correspondence between xls name and xls label
}
\description{
Get a codebook from a dataset with a double header
}
