% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crosstabs_formatted.R
\name{get_crosstabs_formatted}
\alias{get_crosstabs_formatted}
\title{Formatting a set of crosstabulations}
\usage{
get_crosstabs_formatted(
  crosstabs,
  questions,
  choices,
  sep = "/",
  labels = FALSE,
  lang = "English"
)
}
\arguments{
\item{crosstabs}{dataframe; a set of crosstabulations generated with \code{get_crosstabs()} or  \code{get_weighted_crosstabs()}}

\item{questions}{dataframe; the "survey" sheet from an xls-form}

\item{choices}{dataframe; the "choices" sheet from an xls-form}

\item{sep}{character; a separator of variable name and choice name for "select_multiple" sets of variables}

\item{labels}{logical; if a dataframe uses xls value labels instead of value names}

\item{lang}{character; language of labels from xls-form}
}
\value{
dataframe; a set of crosstabulations with variable and value names
}
\description{
Formatting a set of crosstabulations
}
