% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels_names_swaped.R
\name{get_labels_names_swaped}
\alias{get_labels_names_swaped}
\title{Swaping dataframe's header from labels to names and vice versa}
\usage{
get_labels_names_swaped(
  ds,
  codebook,
  name_col = "name",
  label_col = "label",
  result = "names",
  double_header = FALSE
)
}
\arguments{
\item{ds}{dataframe;}

\item{codebook}{dataframe; contains column with xls names and matching xls labels}

\item{name_col}{character; a column from the codebook containing xls names}

\item{label_col}{character;a column from the codebook containing xls labels}

\item{result}{character; target result: "names" or "labels"}

\item{double_header}{logical; if the dataframe has value labels in its first row}
}
\value{
dataframe
}
\description{
Swaping dataframe's header from labels to names and vice versa
}
