% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logic_applied.R
\name{get_logic_applied}
\alias{get_logic_applied}
\title{Applying corrections to the survey logic}
\usage{
get_logic_applied(ds, logic_frame_issues, idvar = "_uuid")
}
\arguments{
\item{ds}{dataframe}

\item{logic_frame_issues}{dataframe; reviewed survey logic issues frame (the issues frame is generated with \code{get_logic_errors_categorical()} or \code{get_logic_errors_categorical_numeric()})}

\item{idvar}{character; a name of unique observation ID (default "_uuid")}
}
\value{
dataframe; updated dataset
}
\description{
Applying corrections to the survey logic
}
