% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logic_check_template.R
\name{get_logic_check_template}
\alias{get_logic_check_template}
\title{Get a template for creating custom logic check}
\usage{
get_logic_check_template()
}
\value{
dataframe; logic check template
}
\description{
Get a template for creating custom logic check
}
