% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names_for_crosstabs.R
\name{get_names_for_crosstabs}
\alias{get_names_for_crosstabs}
\title{Get a vector with dependent variables' names for crosstabulations}
\usage{
get_names_for_crosstabs(
  ds,
  questions,
  ind_var,
  labels = F,
  lang = "English",
  exclude = c(),
  tech = "_"
)
}
\arguments{
\item{ds}{dataframe;}

\item{questions}{dataframe; "survey" sheet of xls-form}

\item{ind_var}{character; name of variable containing unique observation IDs}

\item{labels}{logical; if dataframe's headers are xls-labels instead of names}

\item{lang}{character; label's language (if labels are used instead of names)}

\item{exclude}{character; vector of variables' names that should be excluded from the list of dependent variables}

\item{tech}{character; prefix for technical variables ("_" by default)}
}
\value{
character vector
}
\description{
Get a vector with dependent variables' names for crosstabulations
}
