% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names_for_numerics.R
\name{get_names_for_numerics}
\alias{get_names_for_numerics}
\title{Get a vector of names of numeric variables}
\usage{
get_names_for_numerics(questions, labels = F, lang = "English")
}
\arguments{
\item{questions}{dataframe; "survey" sheet of xls-form}

\item{labels}{logical; if dataframe's headers are xls-labels instead of names}

\item{lang}{character; label's language (if labels are used instead of names)}
}
\value{
character vector
}
\description{
Get a vector of names of numeric variables
}
