% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nas_imputed.R
\name{get_nas_imputed}
\alias{get_nas_imputed}
\title{Impute NAs with a median in numeric variables}
\usage{
get_nas_imputed(
  ds,
  questions,
  idvar = "_uuid",
  labels = F,
  lang = "English",
  exclude = c()
)
}
\arguments{
\item{ds}{dataframe}

\item{questions}{dataframe; "survey" sheet of xls-form}

\item{idvar}{character; name of unique observations IDs variable}

\item{labels}{logical; if dataframe's headers are xls-labels instead of names}

\item{lang}{character; label's language (if labels are used instead of names)}

\item{exclude}{character; vector of variables' names that should not be imputed}
}
\description{
Impute NAs with a median in numeric variables
}
