% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numerics_formatted.R
\name{get_numerics_formatted}
\alias{get_numerics_formatted}
\title{Add xls labels to the set of numeric summaries}
\usage{
get_numerics_formatted(numerics_tab, questions, lang = "English")
}
\arguments{
\item{numerics_tab}{dataframe; set of summary statistics obtained with \code{get_numerics_summaries()} or \code{get_weighted_numerrics_summaries()}}

\item{questions}{dataframe; "survey" sheet of xls-form}

\item{lang}{character; xls-label's language (if labels are used instead of names)}
}
\value{
dataframe
}
\description{
Add xls labels to the set of numeric summaries
}
