% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numerics_summaries.R
\name{get_numerics_summaries}
\alias{get_numerics_summaries}
\title{Set of summary statistics for numeric variables}
\usage{
get_numerics_summaries(ds, numeric_vars, ind_var, wide = T)
}
\arguments{
\item{ds}{dataframe}

\item{numeric_vars}{character; vector of numeric variables' names}

\item{ind_var}{character; name of independent variable}

\item{wide}{logical; "wide-formatted" data}
}
\value{
dataframe
}
\description{
Set of summary statistics for numeric variables
}
