% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_others_extracted.R
\name{get_others_extracted}
\alias{get_others_extracted}
\title{Extract open-ended "other" responses from a survey dataset}
\usage{
get_others_extracted(
  ds,
  idvar = "_uuid",
  suffix = "_other",
  questions,
  choices,
  lang = "English"
)
}
\arguments{
\item{ds}{dataframe;}

\item{idvar}{character; a name of unique observation ID (default "_uuid")}

\item{suffix}{character; suffix that differentiates a variable name, containing open-ended specification of other value from the variable, containing the general question}

\item{questions}{dataframe; "survey" sheet of xls-form}

\item{choices}{dataframe; "choices" sheet of xls-form}

\item{lang}{character; xls-label's language (if labels are used instead of names)}
}
\value{
dataframe
}
\description{
Extract open-ended "other" responses from a survey dataset
}
