% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outliers_med_sd.R
\name{get_outliers_med_sd}
\alias{get_outliers_med_sd}
\title{Exptract values beyond menian +- 3sd}
\usage{
get_outliers_med_sd(
  ds,
  questions,
  idvar = "_uuid",
  labels = F,
  lang = "English",
  exclude = c(),
  lower = F
)
}
\arguments{
\item{ds}{dataframe}

\item{questions}{dataframe; "survey" sheet of xls-form}

\item{idvar}{character; a name of unique observation ID (default "_uuid")}

\item{labels}{logical; if dataframe's headers are xls-labels instead of names}

\item{lang}{character; xls-label's language (if labels are used instead of names)}

\item{exclude}{character; variable names to be excluded from imputation}

\item{lower}{logical; not only +3sd but -3sd outliers are tracked}
}
\description{
Exptract values beyond menian +- 3sd
}
