% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outliers_replaced.R
\name{get_outliers_replaced}
\alias{get_outliers_replaced}
\title{Replace outliers in a dataset with user-defned values}
\usage{
get_outliers_replaced(outliers, ds, idvar = "_uuid")
}
\arguments{
\item{outliers}{dataframe; manually edited dataframe with outliers (initially generated with \code{get_outliers_med_sd()})}

\item{ds}{dataframe;}

\item{idvar}{character; a name of unique observation ID (default "_uuid")}
}
\value{
dataframe
}
\description{
Replace outliers in a dataset with user-defned values
}
