% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppt_graphs_categorical.R
\name{get_ppt_graphs_categorical}
\alias{get_ppt_graphs_categorical}
\title{Print bar charts for categorical variables to Power Point document}
\usage{
get_ppt_graphs_categorical(
  cr_tabs_g,
  ppt_file,
  ppt_layout,
  ppt_theme,
  indicator_col = "indicator",
  question_label = "label::English",
  choice_label = "choice_label",
  unit_col = "unit",
  value_col = "prop",
  type_col = "type"
)
}
\arguments{
\item{cr_tabs_g}{dataset; set of crosstabulations in a long format (generated with \code{get_crosstabs()} or \code{get_weighted_crosstabs()}, "wide" parameter should be set to FALSE) with variables' and choices' labels added (could be done with \code{get_crosstabs_formatted()})}

\item{ppt_file}{character; name of the target ppt file (with an extension)}

\item{ppt_layout}{character; name of a layout in a ppt file (suggested option - title and content)}

\item{ppt_theme}{character; name of a theme in a ppt file}

\item{indicator_col}{character; name of a variable, containing indicators}

\item{question_label}{character; name of a variable, containing questions' labels}

\item{choice_label}{character; name of a variable, containing choices' labels}

\item{unit_col}{character; name of a variable, containing a unit of disaggregation (independent variable)}

\item{value_col}{character; name of a variable, containing numeric values of proportions}

\item{type_col}{character; name of a variable, containing a question type (select_one or select_multiple)}
}
\value{
prints to the ppt file as a side effect
}
\description{
Print bar charts for categorical variables to Power Point document
}
