% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recoded_into_ds.R
\name{get_recoded_into_ds}
\alias{get_recoded_into_ds}
\title{Return recoded "other" responses to a dataset}
\usage{
get_recoded_into_ds(others_ds_done, ds, questions, idvar = "_uuid", sep = ".")
}
\arguments{
\item{others_ds_done}{dataframe; manually reviewed and recoded set of "other" responses}

\item{ds}{dataframe;}

\item{questions}{dataframe; "survey" sheet of xls-form}

\item{idvar}{character; a name of unique observation ID (default "_uuid")}

\item{sep}{character; separator used in "select multiple" questions}
}
\value{
dataframe
}
\description{
Return recoded "other" responses to a dataset
}
