% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sm_delabelled.R
\name{get_sm_delabelled}
\alias{get_sm_delabelled}
\title{Remove choice labels from "select multiple" questions}
\usage{
get_sm_delabelled(ds, sep, label_true, label_false)
}
\arguments{
\item{ds}{dataframe;}

\item{sep}{character; separator used in "select multiple" questions}

\item{label_true}{character; label, used for "true"}

\item{label_false}{character; label used for "false"}
}
\value{
dataframe;
}
\description{
Remove choice labels from "select multiple" questions
}
