% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_t_mean_group.R
\name{get_t_mean_groups}
\alias{get_t_mean_groups}
\title{One-sample t-test for numeric variables}
\usage{
get_t_mean_groups(num_vars, group, ds)
}
\arguments{
\item{num_vars}{character; vector of names of numeric variables to be tested (can be generated with \code{get_names_for_numerics()})}

\item{group}{character; name of an inependent variable}

\item{ds}{dataframe}
}
\value{
dataframe
}
\description{
One-sample t-test for numeric variables
}
