% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weighted_crosstabs.R
\name{get_weighted_crosstabs}
\alias{get_weighted_crosstabs}
\title{Set of weighted crosstabulations from a dataset}
\usage{
get_weighted_crosstabs(ds, ind_var, dep_vars, weights, wide = T)
}
\arguments{
\item{ds}{dataframe;}

\item{ind_var}{character; name of an independent variable}

\item{dep_vars}{character vector; names of dependent variables (can be generated with \code{get_names_for_crosstabs()})}

\item{weights}{character; name of the weighting variable}

\item{wide}{logical; "wide-formatted" data}
}
\value{
dataframe; a set of crosstabulations
}
\description{
Set of weighted crosstabulations from a dataset
}
