% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weighted_numerics.R
\name{get_weighted_numerics}
\alias{get_weighted_numerics}
\title{Set of weighted summary statistics for numeric variables}
\usage{
get_weighted_numerics(ds, numeric_vars, ind_var, weights)
}
\arguments{
\item{ds}{dataframe}

\item{numeric_vars}{character; vector of numeric variables' names}

\item{ind_var}{character; name of independent variable}

\item{weights}{character; name of the weighting variable}
}
\value{
dataframe
}
\description{
Set of weighted summary statistics for numeric variables
}
